fixes #5 - adds 'test-enhance' and 'test-enhance-check' goals; (rebased on top of 5.0, with improvements after Andy's review)
