+++ .gitignore
@@ -1,2 +1,3 @@
+*.iml
 /bin
 /target
