+++ src/main/java/org/datanucleus/maven/SchemaToolCreateMojo.java
@@ -33,6 +33,18 @@ public class SchemaToolCreateMojo extends AbstractSchemaToolMojo
     private static final String OPERATION_MODE_CREATE = "-create";
 
     /**
+     * @parameter expression="${classpath}" default-value="${project.compileClasspathElements}"
+     * @required
+     */
+    private List classpathElements;
+
+    @Override
+    List getClasspathElements() {
+        return classpathElements;
+    }
+
+
+    /**
      * {@inheritDoc}
      * @see org.datanucleus.maven.AbstractSchemaToolMojo#prepareModeSpecificCommandLineArguments(org.codehaus.plexus.util.cli.Commandline, java.util.List)
      */
