+++ src/main/java/org/datanucleus/maven/SchemaToolDeleteDatabaseMojo.java
@@ -32,6 +32,19 @@ public class SchemaToolDeleteDatabaseMojo extends AbstractSchemaToolMojo
     private static final String OPERATION_MODE_DELETE = "-deleteDatabase";
 
     /**
+     * @parameter expression="${classpath}" default-value="${project.compileClasspathElements}"
+     * @required
+     */
+    private List classpathElements;
+
+    @Override
+    List getClasspathElements() {
+        return classpathElements;
+    }
+
+
+
+    /**
      * {@inheritDoc}
      * @see org.datanucleus.maven.AbstractSchemaToolMojo#prepareModeSpecificCommandLineArguments(org.codehaus.plexus.util.cli.Commandline, java.util.List)
      */
