+++ src/main/java/org/datanucleus/maven/SchemaToolDeleteMojo.java
@@ -32,6 +32,19 @@ public class SchemaToolDeleteMojo extends AbstractSchemaToolMojo
     private static final String OPERATION_MODE_DELETE = "-delete";
 
     /**
+     * @parameter expression="${classpath}" default-value="${project.compileClasspathElements}"
+     * @required
+     */
+    private List classpathElements;
+
+    @Override
+    List getClasspathElements() {
+        return classpathElements;
+    }
+
+
+
+    /**
      * {@inheritDoc}
      * 
      * @see org.datanucleus.maven.AbstractSchemaToolMojo#prepareModeSpecificCommandLineArguments(org.codehaus.plexus.util.cli.Commandline, java.util.List)
